Changed sql query to ORM

+++ services/product.service.js
@@ -1,8 +1,10 @@
 import ProductRepository from "../repositories/product.repository.js"
 import SupplierRepository from "../repositories/supplier.repository.js"
+import SaleRepository from "../repositories/sale.repository.js"
+import { Error } from "sequelize";
 
 async function createProduct(product){
-    if(await SupplierRepository.getSupplier(product.supplier_id)) {
+    if(await SupplierRepository.getSupplier(product.supplierId)) {
         return await ProductRepository.insertProduct(product);
     }
     throw new Error("Supplier ID informado não existe!")    
@@ -17,13 +19,17 @@ async function getProduct(id){
 }
 
 async function updateProduct(product){
-    if(await SupplierRepository.getSupplier(product.supplier_id)) {
+    if(await SupplierRepository.getSupplier(product.supplierId)) {
         return await ProductRepository.updateProduct(product);
     }
     throw new Error("Supplier ID informado não existe!")  
 }
 
 async function deleteProduct(id){
+    const sales = await SupplierRepository.getSupplier(product.supplierId);
+    if( sales ) {
+        throw new Error("Não é possível excluir, pois já houve vendas com este produto!")
+    }
     await ProductRepository.deleteProduct(id);
 }
